// src/enums/main.d.ts
export declare enum AppDirection {
    UP = 0,
    DOWN = 1,
    LEFT = 2,
    RIGHT = 3
}
export declare enum TaskStatus {
    PENDING = 1,
    IN_PROGRESS = 5,
    DONE = 10
}
export declare enum UserRole {
    ADMIN = "ADMIN",
    USER = "USER",
    GUEST = "GUEST"
}
export declare enum ResponseFlag {
    YES = 1,
    NO = "NO"
}
export declare enum DeliveryStatus {
    PENDING = "PENDING",
    SHIPPED = "SHIPPED",
    DELIVERED = "DELIVERED"
}
//# sourceMappingURL=main.d.ts.map